move record accessor completion into Ide.Completion
